Add Nascent Chaos proc timeouts to Infuriate
